completion till 25Aug2023

+++ ToDoApp/routes/todos.py
@@ -4,7 +4,7 @@ from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
 from models import ToDo
-
+from .auth import get_current_user
 
 
 
@@ -27,16 +27,25 @@ class ToDoRequest(BaseModel):
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
-
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 @router.get("/")
-async def get_all_todos(db: db_dependency):
-    return db.query(ToDo).all()
+async def get_all_todos(user: user_dependency, db: db_dependency):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    return db.query(ToDo).filter(ToDo.ownerid == user.get('userid')).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def get_single_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    fetched_todo = db.query(ToDo).filter(ToDo.id == todo_id).first()
+async def get_single_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    fetched_todo = db.query(ToDo).filter(ToDo.id == todo_id)\
+                                 .filter(ToDo.ownerid == user.get('userid')).first()
 
     if fetched_todo is not None:
         return fetched_todo
@@ -45,16 +54,28 @@ async def get_single_todo(db: db_dependency, todo_id: int = Path(gt=0)):
 
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, new_todo: ToDoRequest):
-    todo_model = ToDo(**new_todo.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, new_todo: ToDoRequest):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    todo_model = ToDo(**new_todo.model_dump(), ownerid=user.get('userid'))
     db.add(todo_model)
     db.commit()
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, updated_todo: ToDoRequest, todo_id: int = Path(gt=0)):
+async def update_todo(db: db_dependency, 
+                      user: user_dependency,
+                      updated_todo: ToDoRequest, 
+                      todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
     todo_model = ToDo(**updated_todo.model_dump())
-    todo_db = db.query(ToDo).filter(ToDo.id == todo_id).first()
+    todo_db = db.query(ToDo).filter(ToDo.id == todo_id)\
+                .filter(ToDo.ownerid == user.get('userid')).first()
 
     todo_db.title = todo_model.title
     todo_db.description = todo_model.description
@@ -74,11 +95,17 @@ async def update_todo(db: db_dependency, updated_todo: ToDoRequest, todo_id: int
     """
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_to_delete = db.query(ToDo).filter(ToDo.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed.")
+
+    todo_to_delete = db.query(ToDo).filter(ToDo.id == todo_id)\
+                        .filter(ToDo.ownerid == user.get('userid')).first()
+
     if todo_to_delete is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="ToDo not found.")
     
-    db.query(ToDo).filter(ToDo.id == todo_id).delete()
+    db.query(ToDo).filter(ToDo.id == todo_id).filter(ToDo.ownerid == user.get('userid')).delete()
     db.commit()
